added util to parse YAML properties (only moved methods from AbstractSubSwitch to PropertiesParserUtil)
